add append_reference_to_input toggle to isnvs_vcf
add `append_reference_to_input` toggle to `tasks_intrahost.isnvs_vcf` to allow single-segment reference sequences to be added to input fasta, useful for MSAs where the reference has been removed, or where the input fastas are already in the same coordinate space as the reference (as in the case of aligned-to-ref). Defaults to false. `append_reference_to_input` is set to `true` in the `assemble_refbased` workflow to allow for a bypass of the usual mafft alignment step, since we know the output from the align-to-ref task is already in the correct coordinate space.