use WDL value for submitting_lab_name in gisaid_meta_prep python body assertion
The python body of the gisaid_meta_prep task includes an assertion to verify the submitting_lab_name value is in the address map json file, and emits an error if not. This fixes an issue where the error text python f-string fails to evaluate because it is looking for the variable in python scope rather than using a string pre-substituted by the WDL executor before running the python block.