allow docker image pinning to be skipped with #skip-global-version-pin annotation
The script `github_actions_ci/version-wdl-runtimes.sh` is used to replace docker image tags in WDL task files with the tags listed in the `requirements-modules.txt` file. These replacements were previously applied globally. With this commit, image references on lines ending with the comment `#skip-global-version-pin` are exempt from these replacements. This allows tasks to have image versions held back; this is useful in the event that a task has commands that are incompatible with newer image versions.