expose numberOfNegativeControls in illumina_demux (#498)
In the event demux is run with a samplesheet containing only one sample row, find_uncertain_barcodes() will fail since it assumes at least one NTC is present in the samplesheet. The changes in this commit allow the user to set this param to zero, which should allow outlier detection to succeed.