In align_and_count, change default to count only properly-paired reads
Change the input, `filter_bam_to_proper_primary_mapped_reads`, in `align_and_count` to a default value of `true`; this changes the behavior of the task. Before filtering was available as an optional operation in `align_and_count`, reads were not filtered to proper pairs and later on the value of `filter_bam_to_proper_primary_mapped_reads=false` maintained the no-filter approach. With this commit, only properly-paired mapped reads are counted in `align_and_count`.