add workflow to download multiple SRA accessions to multiple bams: fetch_multiple_sra_to_bams
add workflow to download multiple SRA accessions to multiple bams: fetch_multiple_sra_to_bams; this is useful in the event a sample is associated with multiple sequencing runs (i.e. more than one SRR###). It also adjusts the Fetch_SRA_to_BAM task to find metadata for the requested run, in the event multi-run metadata is returned for a given accession