add workflow "unpack_archive_to_bucket"; add task "unpack_archive_to_bucket_path" (#573)
* add workflow "unpack_archive_to_bucket"; add task "unpack_archive_to_bucket_path"

This adds a new WDL task, "unpack_archive_to_bucket_path", to unpack an input tarball to a specified bucket path, with options to set the target destination, layers of wrapping directories to remove (from around the tarball contents), and whether the files extracted should be uploaded directly via pipe or from a temporary location on-disk. This task is intended to run on Terra, but it can be run elsewhere if a Google Cloud auth token is passed in to the task. A corresponding workflow, unpack_archive_to_bucket, is also added.

* pass in gcloud access token, if provided, to unpack_archive_to_bucket_path; otherwise, obtain it via `gcloud auth print-access-token`

pass in gcloud access token, if provided, to unpack_archive_to_bucket_path; otherwise, obtain it via `gcloud auth print-access-token` (which should work on Terra/GCE instances).

* unpack_archive_to_bucket_path: allow additional opts to be passed to `gcloud storage cp`; reorder of inputs and their descriptions

* update unpack_archive_to_bucket_path to viral-core 2.4.1

* add fc- prefix to bucket ID returned by check_terra_env

* additional input descriptive text