add collapse_duplicated_barcodes param to illumina_demux
add `collapse_duplicated_barcodes` input param to `illumina_demux` task, which will be passed to `illumina.py illumina_demux`, and will toggle collapsing of duplicated barcodes or barcode pairs, allowing picard to run IlluminaBaseCallsToSam without the errors that would normally be raised if the sample sheet provided has dups. This new param is set to false by default, so as to not change expected behavior of pipelines currently using the task.